separate やじるしボタンが呼ぶ関数
